Auto numbering from the selection List
The selection list merely has the minimal display string.  The
enumeration of the selection is done based on its position in the
selections list.  The selections list is a collection of tuples.  The
first member of the tuple is the "display name" for the selection.  The
second member is the target.